// src/api/generated/model/apiResponseListCandidateDto.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Measuring Parent API
 * API documentation for Measuring Parent application
 * OpenAPI spec version: 1.0.0
 */
import type { CandidateDto } from "./candidateDto";

export interface ApiResponseListCandidateDto {
  success?: boolean;
  message?: string;
  data?: CandidateDto[];
  errors?: string[];
  timestamp?: string;
  path?: string;
}