// src/api/generated/model/apiResponseListExamDto.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Measuring Parent API
 * API documentation for Measuring Parent application
 * OpenAPI spec version: 1.0.0
 */
import type { ExamDto } from "./examDto";

export interface ApiResponseListExamDto {
  success?: boolean;
  message?: string;
  data?: ExamDto[];
  errors?: string[];
  timestamp?: string;
  path?: string;
}