// src/api/generated/model/createApplicationRequest.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Measuring Parent API
 * API documentation for Measuring Parent application
 * OpenAPI spec version: 1.0.0
 */

export interface CreateApplicationRequest {
  /**
   * @minLength 0
   * @maxLength 255
   */
  name: string;
  /**
   * @minLength 0
   * @maxLength 50
   */
  code: string;
  /** @minLength 1 */
  examId: string;
  /** @minLength 1 */
  examSessionId: string;
  /** @minLength 1 */
  candidateId: string;
  /**
   * @minLength 0
   * @maxLength 50
   */
  username?: string;
}